requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0
matplotlib>=3.7.0
seaborn>=0.12.0
shap>=0.42.0
pyyaml>=6.0
tqdm>=4.65.0
opacus>=1.4.0
flwr>=1.5.0
tensorboard>=2.13.0
joblib>=1.3.0
